Allow same RET RTP port for multiple channels in sig_mode=STD
When sig_mode=STD, VQEC would assume that the receiving (local) port would be the same as the sending (remote) port, and
would refuse to bind a second tuner if another tuner is already bound to another channel with the same repair port.

This is incorrect, as an ephemeral port will be allocated as the local port in this case, and so there is no conflict.